Add function of playGame()
Added a function to play 5 rounds of the game using a For loop.
The const for humanChoice and computerChoice where moved inside that
loop in order to prompt the choices. Otherwise it just kept the initial
choice and ran the 5 games.
Added a tieScore to track the total number of games played for better
understanding of the iterations.
Added an if comparison to log whoever won the game. It doesn't account
for the ties.